+++ CHANGES.rst
@@ -4,14 +4,18 @@ Release History
 Release 0.3.0.dev
 ~~~~~~~~~~~~~~~~~
 
-Release 0.2.1.dev
+Release 0.2.2.dev
 ~~~~~~~~~~~~~~~~~
+
+Release 0.2.1 (2016-08-18)
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 New Features:
 
 * optionally limit the width of large images and make use of this to simulate
   the Sphinx LaTeX builder behavior (#46)
 * reStructuredText/Sphinx: support for images with hyperlinks (#49)
+* record the styled page numbers in the PDF as page labels (#41)
 * unsupported Python versions: prevent installation where possible (sdist)
   or exit on import (wheel)
 * support Python 3.6
@@ -22,6 +26,7 @@ Bugfixes:
   option can actually be used
 * Fix #47: ClassNotFound exception in Literal_Block.lexer_getter()
 * Fix #45: Images that don't fit are still placed on the page
+* don't warn about duplicate style matches that resolve to the same style
 
 
 Release 0.2.0 (2016-08-10)